Use Correctly Sized Values + More
- Now use uint16_t and uint32_t where necessary, as opposed to "int". There can only be uint16_t number of tokens, uint8_t length for names and uint32_t for text length.
- Fix 64-bit by using correct sizes on P/Invoke barrier.
- Add macro-based debugging, _ABP_DEBUG_OUT
- Add exceptions for invalid options.
- Add "MakeBuild" to .gitignore

+++ ABSoftware.ABParser/Internal/NativeMethods.cs
@@ -14,7 +14,7 @@ namespace ABSoftware.ABParser.Internal
         internal const CallingConvention CALLING_CONVENTION = CallingConvention.Cdecl;
 
         [DllImport(COREDLL, CharSet = CHARSET, CallingConvention = CALLING_CONVENTION)]
-        internal static extern IntPtr InitializeTokens(string[] tokensData, int[] tokenLengths, int numberOfTokens, string[] limitNames, int[] limitNameSizes, int[] limitsPerToken);
+        internal static extern IntPtr InitializeTokens(string[] tokensData, ushort[] tokenLengths, ushort numberOfTokens, string[] limitNames, byte[] limitNameSizes, ushort[] limitsPerToken);
 
         [DllImport(COREDLL, CharSet = CHARSET, CallingConvention = CALLING_CONVENTION)]
         internal static extern IntPtr CreateBaseParser(IntPtr tokenData);
@@ -32,7 +32,7 @@ namespace ABSoftware.ABParser.Internal
         internal static extern void DeleteBaseParser(IntPtr baseParser);
 
         [DllImport(COREDLL, CharSet = CHARSET, CallingConvention = CALLING_CONVENTION)]
-        internal static extern void EnterTokenLimit(IntPtr baseParser, string limitName, int limitNameSize);
+        internal static extern void EnterTokenLimit(IntPtr baseParser, string limitName, byte limitNameSize);
 
         [DllImport(COREDLL, CharSet = CHARSET, CallingConvention = CALLING_CONVENTION)]
         internal static extern void ExitTokenLimit(IntPtr baseParser, int levels);
